Move tests for functional to dedicated directory (#1231)
