+++ test/torchaudio_unittest/functional/torchscript_consistency_cuda_test.py
@@ -1,16 +1,16 @@
 import torch
 
 from torchaudio_unittest.common_utils import skipIfNoCuda, PytorchTestCase
-from .torchscript_consistency_impl import Transforms
+from .torchscript_consistency_impl import Functional
 
 
 @skipIfNoCuda
-class TestTransformsFloat32(Transforms, PytorchTestCase):
+class TestFunctionalFloat32(Functional, PytorchTestCase):
     dtype = torch.float32
     device = torch.device('cuda')
 
 
 @skipIfNoCuda
-class TestTransformsFloat64(Transforms, PytorchTestCase):
+class TestFunctionalFloat64(Functional, PytorchTestCase):
     dtype = torch.float64
     device = torch.device('cuda')
